Fixed all discrepancies related to handling of clientId Moved all clientId related tests to ClientIdTest test suite
