+++ src/Channel.php
@@ -59,14 +59,9 @@ class Channel {
 
         $args = func_get_args();
         $json = '';
-        $authClientId = $this->ably->auth->getClientId();
         
         if ( count($args) == 1 && is_a( $args[0], 'Ably\Models\Message' ) ) { // single Message
             $msg = $args[0];
-
-            if ( !empty( $msg->clientId ) && !empty( $authClientId ) && $msg->clientId != $authClientId) {
-                throw new AblyException( 'Message\'s clientId does not match the clientId of the authorisation token.', 40102, 401 );
-            }
 
             if ( $this->options->encrypted ) {
                 $msg->setCipherParams( $this->options->cipherParams );
@@ -77,10 +72,6 @@ class Channel {
             $jsonArray = array();
 
             foreach ( $args[0] as $msg ) {
-                if ( !empty( $msg->clientId ) && !empty( $authClientId ) && $msg->clientId != $authClientId) {
-                    throw new AblyException( 'Message\'s clientId does not match the clientId of the authorisation token.', 40102, 401 );
-                }
-
                 if ( $this->options->encrypted ) {
                     $msg->setCipherParams( $this->options->cipherParams );
                 }
@@ -102,6 +93,12 @@ class Channel {
             $json = $msg->toJSON();
         } else {
             throw new AblyException( 'Wrong parameters provided, use either Message, array of Messages, or name and data', 40003, 400 );
+        }
+        
+        $authClientId = $this->ably->auth->getClientId();
+        // if the message has a clientId set and we're using token based auth, the clientIds must match unless we're a wildcard client
+        if ( !empty( $msg->clientId ) && !$this->ably->auth->isUsingBasicAuth() && $authClientId != '*' && $msg->clientId != $authClientId) {
+            throw new AblyException( 'Message\'s clientId does not match the clientId of the authorisation token.', 40102, 401 );
         }
 
         $this->ably->post( $this->channelPath . '/messages', $headers = array(), $json );
